Time: 0 ms (100.00%) | Memory: 19.6 MB (93.89%) - LeetSync

+++ 662-maximum-width-of-binary-tree/maximum-width-of-binary-tree.cpp
@@ -3,20 +3,20 @@ public:
     int widthOfBinaryTree(TreeNode* root) {
         if (!root) return 0;
 
-        long long ans = 0;
-        queue<pair<TreeNode*, long long>> q;
+        long ans = 0;
+        queue<pair<TreeNode*, long>> q;
         q.push({root, 0});
 
         while (!q.empty()) {
             int size = q.size();
-            long long minIndex = q.front().second;
-            long long first = 0, last = 0;
+            long minIndex = q.front().second;
+            long first = 0, last = 0;
 
             for (int i = 0; i < size; i++) {
                 auto [node, index] = q.front();
                 q.pop();
 
-                long long currIndex = index - minIndex;
+                long currIndex = index - minIndex;
 
                 if (i == 0) first = currIndex;
                 if (i == size - 1) last = currIndex;
